Cleaned up database referencing.
I looked into using a CursorLoader like the Android documentation
recommends, however it required me to implement ContentProviders
for everything, which would have made me deal with URI's and a
more complicated Manifest.

This is not to argue that Loaders are the wrong way to go. They
sound much more efficient. However, right now everything is pretty
light weight, so hopefully it will be fine.

Someday, it should be switched to a CursorLoader. Write a
ContentProvider that takes the place of each DbAdapter, along with
a base ContentProvider. I understand how the Loader makes it better
when I'm setting the Cursor in something like a ListView (or any
other View that requires an Adapter), but I don't quite understand
the access pattern for things like where I just need to modify the
database.

More research is needed, but this should be safe for now.