WOW IT WORKS!
Yes, I'm a terrible committer. Everything seems to be in a working
state now. The simulator tab finally works. It is based off of a
simple rule:

If any members don't have a loan out AND there's still money left
in the pot by the end of a meeting, pick one of them to loan it out
to and automatically have them repay the loan with interest at the
start of the next meeting (only to the simulator tab of course).

Seems to work really well!